edit cutscene b4_prologue
