+++ Assets/Animations/Anim Camera/Prologue/cameraAnimation.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3ee8089e5b8e29b44bd7c4c253482871
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
