--- Assets/Scenes/WIP level 2.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 74b9c9ba0f26df24985a27b30e906afc
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
