+++ Assets/Scripts/CutScene/Templates.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 488ac4556a0c1c14c8928c44e52fe11e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
