+++ Assets/Scripts/CutScene/Templates/beforePrologue_cutscene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 113d69a2acdabef4ead4c3fe02e88926
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
