--- Assets/Scripts/CutScene/TemplatesScenes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8abfbbcdcc4eb5f4c879e3c168f0a6ad
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
