--- Assets/Scripts/CutScene/TemplatesScenes/endCutScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b2306028558bd0149a2f329ec6a9dac8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
