--- Assets/Scripts/CutScene/TemplatesScenes/startCutScene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 42bfb50485a00d74f9ef1cf30e51ee23
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
