+++ Assets/Scripts/CutScene/WIP/DisableCutSceneCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b802660da9eddd438e856d224aac9ab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
