+++ Assets/Scripts/CutScene/WIP/TemplatesScenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8abfbbcdcc4eb5f4c879e3c168f0a6ad
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
